+++ frontend/src/App.js
@@ -1,8 +1,7 @@
 import { Routes, Route } from "react-router-dom"
 import HomePage from "./page/HomePage"
-import Login from "./page/Login"
-import Register from "./page/Register"
-import { ProfilePage } from "./page/Profile";
+import LoginPage from "./page/LoginPage"
+import RegisterPage from "./page/RegisterPage"
 
 import PoemPage from "./page/PoemPage"
 import FeedingPage from "./page/FeedingPage"
@@ -12,6 +11,7 @@ import { HomeLayout } from "./components/HomeLayout";
 import  BasicSelect  from "./components/SelectChild";
 import 'antd/dist/antd.min.css'
 import './APP.css'
+import {ManageChildPage} from "./page/ManageChildPage";
 
 function App() {
 
@@ -19,12 +19,11 @@ function App() {
         <Routes>
             <Route path="/" element={<HomeLayout />}>
                 <Route path="/" element={<HomePage />}></Route>
-                <Route path="/login" element={<Login />}></Route>
-                <Route path="/register" element={<Register />}></Route>
+                <Route path="/login" element={<LoginPage />}></Route>
+                <Route path="/register" element={<RegisterPage />}></Route>
             </Route>
             <Route path="/dashboard" element={<ProtectedLayout />}>
-                {/*<Route path="/" element={<Home />}></Route>*/}
-                <Route path="profile" element={<ProfilePage />} />
+                <Route path="manageChild" element={<ManageChildPage />} />
                 <Route path="feeding" element={<FeedingPage />}></Route>
                 <Route path="poemList" element={<PoemPage />}></Route>
                 <Route path="test" element={<BasicSelect />} />
