+++ frontend/src/page/ManageChildPage.js
@@ -1,0 +1,4 @@
+
+export const ManageChildPage = () => {
+    return ("add child here");
+};
