--- frontend/src/page/Profile.jsx
@@ -1,6 +1,0 @@
-import { BasicPage } from "../components/BasicPage";
-import Person from "@mui/icons-material/Person";
-
-export const ProfilePage = () => {
-    return <BasicPage title="Profile Page" icon={<Person />} />;
-};
