+++ src/main/java/com/ssw555/babycare/controller/ChildController.java
@@ -19,8 +19,7 @@ public class ChildController {
     @Autowired
     private ChildRepository childRepository;
 
-    @Autowired
-    private PoemRepository poemRepository;
+
